requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
requests==2.31.0
